Valid.js Project
Valid.js is a javascript file containing functions that can be used by
web applications to validate their application's input at the client
side. Simply by including the file link within the head section of the
web page and using the desired functions. Procedure is mentioned in the
comment section of javascript file.